refactor(product): 테스트 이전
