+++ product/src/test/java/kitchenpos/product/group/MenuGroupAcceptanceTest.java
@@ -1,15 +1,16 @@
-package kitchenpos.product;
+package kitchenpos.product.group;
 
-import static kitchenpos.product.step.MenuGroupAcceptanceStep.메뉴_그룹_등록_되어_있음;
-import static kitchenpos.product.step.MenuGroupAcceptanceStep.메뉴_그룹_등록_됨;
-import static kitchenpos.product.step.MenuGroupAcceptanceStep.메뉴_그룹_등록_요청;
-import static kitchenpos.product.step.MenuGroupAcceptanceStep.메뉴_그룹_목록_조회_됨;
-import static kitchenpos.product.step.MenuGroupAcceptanceStep.메뉴_그룹_목록_조회_요청;
+
+import static kitchenpos.product.group.step.MenuGroupAcceptanceStep.메뉴_그룹_등록_되어_있음;
+import static kitchenpos.product.group.step.MenuGroupAcceptanceStep.메뉴_그룹_등록_됨;
+import static kitchenpos.product.group.step.MenuGroupAcceptanceStep.메뉴_그룹_등록_요청;
+import static kitchenpos.product.group.step.MenuGroupAcceptanceStep.메뉴_그룹_목록_조회_됨;
+import static kitchenpos.product.group.step.MenuGroupAcceptanceStep.메뉴_그룹_목록_조회_요청;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import kitchenpos.AcceptanceTest;
-import kitchenpos.product.ui.response.MenuGroupResponse;
+import kitchenpos.product.AcceptanceTest;
+import kitchenpos.product.group.ui.response.MenuGroupResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
