+++ product/src/test/java/kitchenpos/product/menu/MenuAcceptanceTest.java
@@ -1,20 +1,21 @@
-package kitchenpos.product;
+package kitchenpos.product.menu;
 
-import static kitchenpos.product.step.MenuAcceptanceStep.메뉴_등록_되어_있음;
-import static kitchenpos.product.step.MenuAcceptanceStep.메뉴_등록_됨;
-import static kitchenpos.product.step.MenuAcceptanceStep.메뉴_등록_요청;
-import static kitchenpos.product.step.MenuAcceptanceStep.메뉴_목록_조회_됨;
-import static kitchenpos.product.step.MenuAcceptanceStep.메뉴_목록_조회_요청;
-import static kitchenpos.product.step.MenuGroupAcceptanceStep.메뉴_그룹_등록_되어_있음;
-import static kitchenpos.product.step.ProductAcceptanceStep.상품_등록_되어_있음;
+
+import static kitchenpos.product.group.step.MenuGroupAcceptanceStep.메뉴_그룹_등록_되어_있음;
+import static kitchenpos.product.menu.step.MenuAcceptanceStep.메뉴_등록_되어_있음;
+import static kitchenpos.product.menu.step.MenuAcceptanceStep.메뉴_등록_됨;
+import static kitchenpos.product.menu.step.MenuAcceptanceStep.메뉴_등록_요청;
+import static kitchenpos.product.menu.step.MenuAcceptanceStep.메뉴_목록_조회_됨;
+import static kitchenpos.product.menu.step.MenuAcceptanceStep.메뉴_목록_조회_요청;
+import static kitchenpos.product.product.step.ProductAcceptanceStep.상품_등록_되어_있음;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.math.BigDecimal;
-import kitchenpos.AcceptanceTest;
-import kitchenpos.product.ui.response.MenuGroupResponse;
-import kitchenpos.product.ui.response.MenuResponse;
-import kitchenpos.product.ui.response.ProductResponse;
+import kitchenpos.product.AcceptanceTest;
+import kitchenpos.product.group.ui.response.MenuGroupResponse;
+import kitchenpos.product.menu.ui.response.MenuResponse;
+import kitchenpos.product.product.ui.response.ProductResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
