+++ product/src/test/java/kitchenpos/product/menu/application/MenuServiceTest.java
@@ -1,8 +1,8 @@
-package kitchenpos.product.application;
+package kitchenpos.product.menu.application;
 
-import static kitchenpos.product.sample.MenuGroupSample.두마리메뉴;
-import static kitchenpos.product.sample.MenuSample.이십원_후라이드치킨_두마리세트;
-import static kitchenpos.product.sample.ProductSample.십원치킨;
+import static kitchenpos.product.group.sample.MenuGroupSample.두마리메뉴;
+import static kitchenpos.product.menu.sample.MenuSample.이십원_후라이드치킨_두마리세트;
+import static kitchenpos.product.product.sample.ProductSample.십원치킨;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -13,16 +13,16 @@ import static org.mockito.Mockito.when;
 
 import java.math.BigDecimal;
 import java.util.Collections;
-import kitchenpos.common.domain.Quantity;
-import kitchenpos.product.domain.Menu;
-import kitchenpos.product.domain.MenuGroup;
-import kitchenpos.product.domain.MenuGroupRepository;
-import kitchenpos.product.domain.MenuProduct;
-import kitchenpos.product.domain.MenuRepository;
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductRepository;
-import kitchenpos.product.ui.request.MenuProductRequest;
-import kitchenpos.product.ui.request.MenuRequest;
+import kichenpos.common.domain.Quantity;
+import kitchenpos.product.group.domain.MenuGroup;
+import kitchenpos.product.group.domain.MenuGroupRepository;
+import kitchenpos.product.menu.domain.Menu;
+import kitchenpos.product.menu.domain.MenuProduct;
+import kitchenpos.product.menu.domain.MenuRepository;
+import kitchenpos.product.menu.ui.request.MenuProductRequest;
+import kitchenpos.product.menu.ui.request.MenuRequest;
+import kitchenpos.product.product.domain.Product;
+import kitchenpos.product.product.domain.ProductRepository;
 import org.assertj.core.api.ThrowableAssert.ThrowingCallable;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
