+++ product/src/test/resources/truncate.sql
@@ -1,0 +1,8 @@
+SET FOREIGN_KEY_CHECKS = false;
+
+TRUNCATE TABLE menu_product RESTART IDENTITY;
+TRUNCATE TABLE product RESTART IDENTITY;
+TRUNCATE TABLE menu RESTART IDENTITY;
+TRUNCATE TABLE menu_group RESTART IDENTITY;
+
+SET FOREIGN_KEY_CHECKS = true;
